refactor(cli): split commands.rs into modules
Split crates/citum-cli/src/commands.rs (2923 lines) into a commands/
submodule hierarchy. Per-family dispatch() functions collapse the
top-level run() from 95 lines to a 23-line match.

Small DRY cleanups bundled in:
  - introduce CliResult<T = ()> type alias
  - dedupe style_catalog_row + installed_style_catalog_row into
    StyleCatalogRow::from_entry / StyleCatalogRow::installed
  - centralize confirm() and validate_resource_name() in util.rs

Tests co-located with the code they exercise. 1298/1298 workspace
tests pass; clippy passes under --workspace --all-targets. No
behavior change. Closes item #3 on csl26-rfct.
